+++ changelog.md
@@ -1,3 +1,7 @@
-## Changes for 12.0
-
-* Fixed bugs when using emulate copy in applications like LibreOffice Writer.
+## Changes for 13.0 
+* Fixed an issue in visual layout of the settings panel, thanks to Cyrille Bougot.
+* Improved documentation.
+* Added a Clip Contents Designer category to assign input gestures to all commands available for this add-on.
+* Fixed bugs when using emulate copy in browsers if focus mode is active.
+* You can assign different gestures to show the clipboard textual contents as raw text or formatted in HTML. The Format to show the clipboard text in the settings panel has being modified accordingly, to select the two options available for HTML format.
+* Updated translations.